requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
scipy>=1.10.0
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=1.0.0
tqdm>=4.65.0
